cleanup: block types router
- move helper function elsewhere
